Split Blazor WASM Auth Tests to New Project
